# BACKEND/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tickets/raise": {
      "post": {
        "operationId": "TicketController_raiseTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Raise a new support ticket (public)",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/validate/{trackingId}": {
      "get": {
        "operationId": "TicketController_validateTrackingId",
        "parameters": [
          {
            "name": "trackingId",
            "required": true,
            "in": "path",
            "description": "Shipment tracking ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Validate a tracking ID before raising a ticket (public)",
        "tags": [
          "Tickets"
        ]
      }
    },
...
    "/tickets/agency/{agencyId}": {
      "get": {
        "operationId": "TicketController_getTicketsByAgency",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get tickets by agency",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/{ticketId}/resolve": {
      "post": {
        "operationId": "TicketController_resolveTicket",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Resolve a ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/{ticketId}/status": {
      "post": {
        "operationId": "TicketController_updateStatus",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update ticket status",
        "tags": [
          "Tickets"
        ]
      }
    },
...
    "/agency/{agencyId}/agents": {
      "get": {
        "operationId": "WorkersController_getAgents",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all agents for an agency",
        "tags": [
          "Workers & Agents"
        ]
      },
      "post": {
        "operationId": "WorkersController_addAgent",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Add a new delivery agent to an agency",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
    "/agency/{agencyId}/agents/{agentId}": {
      "delete": {
        "operationId": "WorkersController_removeAgent",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Remove a delivery agent from an agency",
        "tags": [
          "Workers & Agents"
        ]
      },
      "put": {
        "operationId": "WorkersController_editAgent",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Edit a delivery agent's details",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
    "/agency/{agencyId}/assignAgent": {
      "post": {
        "operationId": "WorkersController_assignAgent",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignAgentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Assign an agent to a shipment",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
    "/agency/{agencyId}/assignments": {
      "get": {
        "operationId": "WorkersController_getAllAssignments",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all agent-shipment assignments for an agency",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
    "/agency/{agencyId}/assignments/{trackingId}": {
      "get": {
        "operationId": "WorkersController_getAssignment",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackingId",
            "required": true,
            "in": "path",
            "description": "Shipment tracking ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get assignment details for a specific shipment",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
    "/drivers": {
      "get": {
        "operationId": "WorkersController_getAllDrivers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all drivers",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
    "/drivers/{routeKey}": {
      "get": {
        "operationId": "WorkersController_getDriverByRoute",
        "parameters": [
          {
            "name": "routeKey",
            "required": true,
            "in": "path",
            "description": "Route key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get driver by route key",
        "tags": [
          "Workers & Agents"
        ]
      },
      "put": {
        "operationId": "WorkersController_setDriver",
        "parameters": [
          {
            "name": "routeKey",
            "required": true,
            "in": "path",
            "description": "Route key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Rajesh Kumar"
                  },
                  "vehicle": {
                    "type": "string",
                    "example": "TN-01-AB-1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Set or update a driver for a route",
        "tags": [
          "Workers & Agents"
        ]
      },
      "delete": {
        "operationId": "WorkersController_deleteDriver",
        "parameters": [
          {
            "name": "routeKey",
            "required": true,
            "in": "path",
            "description": "Route key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete a driver from a route",
        "tags": [
          "Workers & Agents"
        ]
      }
    },
...
  },
  "info": {
    "title": "WARE2DOOR API",
    "description": "## WARE2DOOR Logistics API\n\nEnd-to-end logistics platform covering warehouse management, transit hub scanning, last-mile delivery, RTO lifecycle, and super-user administration.\n\n### Authentication\nProtected endpoints require an **`x-role`** header specifying the caller's role.\n\nClick the **Authorize 🔒** button, then enter one of the allowed values:\n\n| Role Value | Access |\n|---|---|\n| `WAREHOUSE` | Warehouse management endpoints |\n| `TRANSIT_HUB` | Transit hub scan & reports endpoints |\n| `LOCAL_AGENCY` | Agency delivery & agent endpoints |\n| `SUPER_USER` | Full administrative access |\n\n> **Public endpoints** (login, verify-email, track, raise-ticket) do **not** require this header.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role-based access header. Allowed values: WAREHOUSE | TRANSIT_HUB | LOCAL_AGENCY | SUPER_USER"
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "trackingId": {
            "type": "string",
            "example": "TRK-HYD-20240001",
            "description": "Tracking ID of the shipment"
          },
          "category": {
            "type": "string",
            "example": "Delayed Shipment",
            "enum": [
              "Delayed Shipment",
              "Damaged Package",
              "Billing & Invoices",
              "Return to Origin",
              "Address Change",
              "Other Issues"
            ],
            "description": "Category of the support ticket"
          },
          "description": {
            "type": "string",
            "example": "My shipment has been delayed for over 5 days.",
            "description": "Detailed description of the issue (min 10 characters)"
          }
        },
        "required": [
          "trackingId",
          "category",
          "description"
        ]
      },
      "UpdateTicketStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "In Progress",
            "enum": [
              "Open",
              "In Progress",
              "Resolved",
              "Closed"
            ],
            "description": "New status for the ticket"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateAgentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Arjun Sharma",
            "description": "Full name of the delivery agent"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Valid 10-digit phone number with optional country code"
          },
          "email": {
            "type": "string",
            "example": "arjun@agency.com",
            "description": "Agent email address (optional)"
          },
          "area": {
            "type": "string",
            "example": "Koramangala, Bangalore",
            "description": "Delivery area assigned to the agent"
          }
        },
        "required": [
          "name",
          "phone",
          "area"
        ]
      },
      "UpdateAgentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Arjun Sharma",
            "description": "Updated name of the agent"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Updated 10-digit phone number"
          },
          "email": {
            "type": "string",
            "example": "arjun@agency.com",
            "description": "Updated email address"
          },
          "area": {
            "type": "string",
            "example": "Indiranagar, Bangalore",
            "description": "Updated delivery area"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Busy",
              "Offline"
            ],
            "description": "Current status of the agent"
          }
        }
      },
      "AssignAgentDto": {
        "type": "object",
        "properties": {
          "trackingId": {
            "type": "string",
            "example": "TRK-HYD-20240001",
            "description": "Tracking ID of the shipment to assign"
          },
          "agentId": {
            "type": "string",
            "example": "AGT-BLR001-1710000000000",
            "description": "ID of the agent to assign"
          }
        },
        "required": [
          "trackingId",
          "agentId"
        ]
      },
...
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}